requirement.txt
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.1.2
html-testRunner==1.1.2
idna==2.7
Jinja2==2.9.5
jsonpath-rw==1.4.0
MarkupSafe==1.0
mock==2.0.0
pbr==4.3.0
ply==3.11
pytz==2018.5
PyYAML==3.13
requests==2.19.1
ruamel.yaml==0.15.75
six==1.11.0
urllib3==1.23
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
